requirements.txt
# Django Core
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5

# JWT Authentication
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# Production Server
gunicorn>=21.2.0
whitenoise>=6.6.0

# Utilities
python-dateutil>=2.8.2
Pillow>=10.0.0
